cmd/llm-inference/pyproject.toml
[project]
name = "llm-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "numexpr>=2.10.2",
    "protobuf>=6.33.1",
    "sentencepiece>=0.2.1",
    "transformers>=4.57.1",
]
